pyproject.toml
[project]
name = "dataset-analyzer"
version = "0.1.0"
description = "Object detection dataset analyzer"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
]
